requirements.txt
pandas
TA-Lib
matplotlib
requests
ccxt
keras
tensorflow
scikit-learn
python-telegram-bot
